build vfd against 17.11 DPDK
Building VFd against 17.11 was failing. This patch fixes that.

+++ src/vfd/vfd_mlx5.c
@@ -9,7 +9,7 @@
 #include <string.h>
 
 int
-vfd_mlx5_get_ifname(uint8_t port_id, char *ifname)
+vfd_mlx5_get_ifname(uint16_t port_id, char *ifname)
 {
 	struct rte_eth_dev_info dev_info;
 	rte_eth_dev_info_get(port_id, &dev_info);
@@ -21,7 +21,7 @@ vfd_mlx5_get_ifname(uint8_t port_id, char *ifname)
 }
 
 int
-vfd_mlx5_get_num_vfs(uint8_t port_id)
+vfd_mlx5_get_num_vfs(uint16_t port_id)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -47,7 +47,7 @@ vfd_mlx5_get_num_vfs(uint8_t port_id)
 }
 
 int
-vfd_mlx5_set_vf_link_status(uint8_t port_id, uint16_t vf_id, int status)
+vfd_mlx5_set_vf_link_status(uint16_t port_id, uint16_t vf_id, int status)
 {
 	char ifname[IF_NAMESIZE];
 	char link_state[16]= "";
@@ -83,7 +83,7 @@ vfd_mlx5_set_vf_link_status(uint8_t port_id, uint16_t vf_id, int status)
 }
 
 int 
-vfd_mlx5_set_vf_mac_addr(uint8_t port_id, uint16_t vf_id, const char* mac)
+vfd_mlx5_set_vf_mac_addr(uint16_t port_id, uint16_t vf_id, const char* mac)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -104,7 +104,7 @@ vfd_mlx5_set_vf_mac_addr(uint8_t port_id, uint16_t vf_id, const char* mac)
 }
 
 int
-vfd_mlx5_vf_mac_remove(uint8_t port_id, uint16_t vf_id)
+vfd_mlx5_vf_mac_remove(uint16_t port_id, uint16_t vf_id)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -125,7 +125,7 @@ vfd_mlx5_vf_mac_remove(uint8_t port_id, uint16_t vf_id)
 }
 
 int 
-vfd_mlx5_set_vf_vlan_stripq(uint8_t port_id, uint16_t vf_id, uint8_t on)
+vfd_mlx5_set_vf_vlan_stripq(uint16_t port_id, uint16_t vf_id, uint8_t on)
 {
 	char ifname[IF_NAMESIZE];
 	char vf_id_s[8] = "";
@@ -143,7 +143,7 @@ vfd_mlx5_set_vf_vlan_stripq(uint8_t port_id, uint16_t vf_id, uint8_t on)
 
 
 int 
-vfd_mlx5_set_vf_vlan_insert(uint8_t port_id, uint16_t vf_id, uint16_t vlan_id)
+vfd_mlx5_set_vf_vlan_insert(uint16_t port_id, uint16_t vf_id, uint16_t vlan_id)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -164,7 +164,7 @@ vfd_mlx5_set_vf_vlan_insert(uint8_t port_id, uint16_t vf_id, uint16_t vlan_id)
 }
 
 int
-vfd_mlx5_set_vf_rate_limit(uint8_t port_id, uint16_t vf_id, uint16_t rate)
+vfd_mlx5_set_vf_rate_limit(uint16_t port_id, uint16_t vf_id, uint16_t rate)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -185,7 +185,7 @@ vfd_mlx5_set_vf_rate_limit(uint8_t port_id, uint16_t vf_id, uint16_t rate)
 }
 
 int
-vfd_mlx5_set_vf_mac_anti_spoof(uint8_t port_id, uint16_t vf_id, uint8_t on)
+vfd_mlx5_set_vf_mac_anti_spoof(uint16_t port_id, uint16_t vf_id, uint8_t on)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -205,7 +205,7 @@ vfd_mlx5_set_vf_mac_anti_spoof(uint8_t port_id, uint16_t vf_id, uint8_t on)
 }
 
 uint32_t
-vfd_mlx5_get_pf_spoof_stats(uint8_t port_id)
+vfd_mlx5_get_pf_spoof_stats(uint16_t port_id)
 {
 	char ifname[IF_NAMESIZE];
 	//FILE *fp;
@@ -263,7 +263,7 @@ vfd_mlx5_get_vf_ethtool_counter(char *ifname, const char *counter)
 }
 
 uint64_t
-vfd_mlx5_get_vf_spoof_stats(uint8_t port_id, uint16_t vf_id)
+vfd_mlx5_get_vf_spoof_stats(uint16_t port_id, uint16_t vf_id)
 {
 	char ifname[IF_NAMESIZE];
 	char tx_err_cnt[64] = "";
@@ -277,7 +277,7 @@ vfd_mlx5_get_vf_spoof_stats(uint8_t port_id, uint16_t vf_id)
 }
 
 int
-vfd_mlx5_get_vf_stats(uint8_t port_id, uint16_t vf_id, struct rte_eth_stats *stats)
+vfd_mlx5_get_vf_stats(uint16_t port_id, uint16_t vf_id, struct rte_eth_stats *stats)
 {
 	char ifname[IF_NAMESIZE];
 	char tx_err_cnt[64] = "";
@@ -323,7 +323,7 @@ vfd_mlx5_pf_vf_offset(char *pciid)
 }
 
 int
-vfd_mlx5_set_prio_trust(uint8_t port_id)
+vfd_mlx5_set_prio_trust(uint16_t port_id)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[128] = "";
@@ -344,7 +344,7 @@ vfd_mlx5_set_prio_trust(uint8_t port_id)
 }
 
 int
-vfd_mlx5_set_qos_pf(uint8_t port_id, sriov_port_t *pf)
+vfd_mlx5_set_qos_pf(uint16_t port_id, sriov_port_t *pf)
 {
 	char ifname[IF_NAMESIZE];
 	char cmd[256] = "";
